requirements.txt
certifi==2022.12.7
charset-normalizer==3.0.1
colorama==0.4.6
etl==0.0.1
filelock==3.9.0
huggingface-hub==0.12.0
idna==3.4
joblib==1.2.0
numpy==1.24.1
packaging==23.0
pandas==1.5.3
python-dateutil==2.8.2
pytz==2022.7.1
PyYAML==6.0
regex==2022.10.31 
requests==2.28.2
scikit-learn==1.2.1
scipy==1.10.0
six==1.16.0
threadpoolctl==3.1.0
tokenizers==0.13.2
torch==1.13.1
tqdm==4.64.1
transformers==4.26.0
typing_extensions==4.4.0
urllib3==1.26.14

transformers == 4.24.0
datasets == 2.7.0
evaluate == 0.3.0
python-box[all]~=6.0
kaggle
gdown